twodgrad to velgrad
