+++ example_channelfit.py
@@ -1,33 +1,34 @@
 from channelfit import ChannelFit
 
 '-------- INPUTS --------'
-cubefits = './channelfit/testcube.fits'
-center = '16h28m21.615s -24d36m23.33s'
-pa = 0  # deg
-incl = 45  # deg
-vsys = 0  # km/s
+cubefits = './testfits/test.cube.fits'
+center = '04h39m53.878s +26d03m09.43s'
+pa = 2.0  # deg
+incl = 85  # deg
+vsys = 5.9  # km/s
 dist = 140  # pc
-sigma = 2e-3  # Jy/beam; None means automatic calculation.
-rmax = 1.5 * dist  # au
-vlim = (-4.0, -0.6, 0.6, 4.0)  # km/s; from vsys
+sigma = 1.7e-3  # Jy/beam; None means automatic calculation.
+rmax = 200  # au
+vlim = (-3.6, -2.0, 2.0, 3.6)  # km/s; from vsys
 '------------------------'
 
 
 '-------- HOW TO DO EACH STEP --------'
-filehead = cubefits.replace('.fits', '')
+filehead = 'test.channelfit'
 chan = ChannelFit(scaling='uniform', progressbar=True)
 chan.makegrid(cubefits=cubefits, center=center, pa=pa, incl=incl,
               vsys=vsys, dist=dist, sigma=sigma, rmax=rmax, vlim=vlim)
 chan.fitting(Mstar_range=[0.1, 2.0],
              Rc_range=[30, 300],
+             h1_range=[0, 0.2],
+             h2_range=[0, 0.2],
              pI_range=[0, 2],
-             fixed_params={'cs': 0.4, 'h1': 0, 'h2': -1,
-                           'Rin': 0, 'Ienv': 0,
+             fixed_params={'cs': 0.4, 'Rin': 0, 'Ienv': 0,
                            'xoff': 0, 'yoff': 0, 'voff': 0,
                            'incloff': 0, 'paoff':0},
              kwargs_emcee_corner={'nwalkers_per_ndim': 4,
-                                  'nburnin': 10,
-                                  'nsteps': 20,
+                                  'nburnin': 100,
+                                  'nsteps': 200,
                                   'rangelevel': 0.99},
              filename=filehead)
 p = chan.popt
