+++ example_linear.py
@@ -2,8 +2,8 @@ import numpy as np
 from pvanalysis import PVAnalysis
 
 '-------- INPUTS --------'
-fitsfile = './testfits/testlinear.fits'
-outname = 'linear'  # file name header for outputs
+fitsfile = './testfits/test.pvanalysis.linear.fits'
+outname = 'test.linear'  # file name header for outputs
 incl = 65.  # deg
 vsys = 4.0  # km/s
 dist = 139.  # pc
