--- twodgrad/__init__.py
@@ -1,3 +1,0 @@
-from ._twodgrad import TwoDGrad
-
-__all__ = ['TwoDGrad']
