+++ velgrad/__init__.py
@@ -1,0 +1,3 @@
+from ._velgrad import VelGrad
+
+__all__ = ['TwoDGrad']
